pyproject.toml
[project]
name = "beetracker"
version = "0.1.0"
description = "Bee tracking software"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.8",
    "opencv-python>=4.13.0.92",
    "openpyxl>=3.1.5",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "scipy>=1.17.1",
    "supabase>=2.28.3",
    "ultralytics>=8.4.37",
]
